moved stray branch into the branches tree

+++ emuv1/msxtypes.h
@@ -7,10 +7,10 @@
 typedef unsigned char nw_byte;          // should be (at least?) an 8 bit type
 typedef unsigned int nw_word;           // should be (at least?) an 16 bit type
     
-typedef unsigned long long emuTimeType;
+typedef unsigned int emuTimeType;
 typedef unsigned long long msTimeType;
 
 #endif
 
-// todo: use SDL types for uint16/uint32 only when interfacing with SDL
-//       and keep the SDL types from spreading throughout isolated classes with dedicated functions.
+// todo: use SDL types for uint16/uint32 through out the emulator
+//       instead of unsigned int or unsigned long
